backend/requirements.txt
aiosqlite
alembic
asyncpg
psycopg2-binary
fastapi
httpx
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pydantic-settings
pytest
pytest-asyncio
python-multipart
sqlalchemy[asyncio]
uvicorn